requirements.txt
django
django-mongodb-backend
voyageai
langchain
langchain-community
python-dotenv
gradio
ipykernel